# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: go-clipboard API
  version: 0.1.0
  description: |
    JSON API for go-clipboard. Go serves only `/api/*`; hit it through the
    Caddy origin in development (`http://localhost:3000`), not the Go port
    directly.

    Product intent lives in `docs/slices/`. This file is the HTTP contract —
    import it into Postman, or copy the examples into curl.

servers:
  - url: http://localhost:3000
    description: Local development (Caddy origin)

paths:
  /api/health:
    get:
      operationId: getHealth
      summary: Stack health
      description: |
        Always returns HTTP 200 when the process can answer. Degraded
        dependencies are reported in the body via `status`, not via the
        status code. The `migrations` object is omitted when the pending
        check could not run (for example when Postgres is unreachable).
      tags:
        - health
      responses:
        "200":
          description: Health payload observed for this request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                ok:
                  summary: Healthy stack
                  value:
                    status: ok
                    database:
                      reachable: true
                      latencyMs: 1
                    migrations:
                      pending: false
                      currentVersion: 1
                    serverTime: "2026-07-27T07:00:00.000000000Z"
                degraded:
                  summary: Database unreachable
                  value:
                    status: degraded
                    database:
                      reachable: false
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/clips:
    post:
      operationId: createClip
      summary: Create a clip
      description: |
        Claim a slug with a text body. Returns 201 when the name is free or
        expired. Live collisions return 409 with `slug_in_use`.
      tags:
        - clips
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClipRequest"
            examples:
              note:
                summary: Simple note (default lifetime)
                value:
                  slug: my-note
                  body: "hello\nworld\n"
              shortLived:
                summary: Ten-minute clip
                value:
                  slug: secret-link
                  body: "temporary password"
                  ttlSeconds: 600
      responses:
        "201":
          description: Clip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateClipResponse"
              example:
                slug: my-note
                expiresAt: "2026-07-27T09:00:00.000000000Z"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Slug is held by a live clip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: slug_in_use
                message: that name is in use right now
        "429":
          $ref: "#/components/responses/RateLimited"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/clips/{slug}:
    get:
      operationId: getClip
      summary: Read a live clip
      description: |
        Returns the stored body for a live clip. Missing and expired slugs both
        return 404 with the same error shape.
      tags:
        - clips
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: Exact slug string (case-sensitive)
      responses:
        "200":
          description: Live clip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadClipResponse"
              example:
                slug: my-note
                body: "hello\nworld\n"
                expiresAt: "2026-07-27T09:00:00.000000000Z"
                serverTime: "2026-07-27T07:00:00.000000000Z"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"

  /api/clips/{slug}/availability:
    get:
      operationId: getClipAvailability
      summary: Check whether a slug looks claimable
      description: |
        Advisory read-only check: reports whether a name currently appears free.
        Does not reserve the name. Invalid or reserved slugs return the same
        validation error shape as create.
      tags:
        - clips
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: Exact slug string (case-sensitive)
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClipAvailabilityResponse"
              examples:
                available:
                  summary: Name looks free
                  value:
                    slug: my-note
                    available: true
                taken:
                  summary: Live clip holds the name
                  value:
                    slug: my-note
                    available: false
        "400":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - database
      properties:
        status:
          type: string
          enum: [ok, degraded]
          description: Overall status derived from the checks that completed
        database:
          $ref: "#/components/schemas/DatabaseHealth"
        migrations:
          $ref: "#/components/schemas/MigrationHealth"
          description: Present only when the pending-migration check completed
        serverTime:
          type: string
          format: date-time
          description: Postgres `now()` as RFC3339Nano UTC, when reachable

    DatabaseHealth:
      type: object
      required:
        - reachable
      properties:
        reachable:
          type: boolean
        latencyMs:
          type: integer
          format: int64
          description: Round-trip latency in milliseconds when reachable

    MigrationHealth:
      type: object
      required:
        - pending
        - currentVersion
      properties:
        pending:
          type: boolean
        currentVersion:
          type: integer
          format: int64

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable message

    CreateClipRequest:
      type: object
      required:
        - slug
        - body
      properties:
        slug:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
          description: Case-sensitive clip name
        body:
          type: string
          maxLength: 262144
          description: UTF-8 text body (256 KiB max)
        ttlSeconds:
          type: integer
          format: int64
          enum: [600, 3600, 7200]
          description: |
            Chosen lifetime in seconds from the anonymous preset set (10 minutes,
            1 hour, or 2 hours). Omitted values use the default preset (2 hours).

    CreateClipResponse:
      type: object
      required:
        - slug
        - expiresAt
      properties:
        slug:
          type: string
        expiresAt:
          type: string
          format: date-time
          description: Expiry anchored to creation time as RFC3339Nano UTC

    ReadClipResponse:
      type: object
      required:
        - slug
        - body
        - expiresAt
        - serverTime
      properties:
        slug:
          type: string
        body:
          type: string
        expiresAt:
          type: string
          format: date-time
        serverTime:
          type: string
          format: date-time
          description: Server clock at response time as RFC3339Nano UTC

    ClipAvailabilityResponse:
      type: object
      required:
        - slug
        - available
      properties:
        slug:
          type: string
        available:
          type: boolean
          description: True when the slug passes validation and has no live clip

  responses:
    NotFound:
      description: Unknown path
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: not_found
            message: resource not found

    MethodNotAllowed:
      description: Path exists but does not answer this method
      headers:
        Allow:
          schema:
            type: string
          description: Comma-separated methods this path accepts
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: method_not_allowed
            message: method not allowed

    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: rate_limited
            message: too many requests

    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidSlug:
              summary: Invalid slug
              value:
                code: invalid_slug
                message: slug must be 3–64 characters
            reservedSlug:
              summary: Reserved slug
              value:
                code: reserved_slug
                message: that name is reserved
            bodyTooLarge:
              summary: Body too large
              value:
                code: body_too_large
                message: clip body must be at most 256 KiB
            invalidBody:
              summary: Invalid JSON body
              value:
                code: invalid_body
                message: request body must be valid JSON with slug and body fields
            invalidTTL:
              summary: Lifetime not in preset set
              value:
                code: invalid_ttl
                message: lifetime must be 10 minutes, 1 hour, or 2 hours