# internal/openapi/openapi.gen.go
//go:build go1.22

// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package openapi

import (
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// Defines values for CreateClipRequestTtlSeconds.
const (
	N3600 CreateClipRequestTtlSeconds = 3600
	N600  CreateClipRequestTtlSeconds = 600
	N7200 CreateClipRequestTtlSeconds = 7200
)

// Valid indicates whether the value is a known member of the CreateClipRequestTtlSeconds enum.
func (e CreateClipRequestTtlSeconds) Valid() bool {
	switch e {
	case N3600:
		return true
	case N600:
		return true
	case N7200:
		return true
	default:
		return false
	}
}

// Defines values for HealthResponseStatus.
const (
	Degraded HealthResponseStatus = "degraded"
	Ok       HealthResponseStatus = "ok"
)

// Valid indicates whether the value is a known member of the HealthResponseStatus enum.
func (e HealthResponseStatus) Valid() bool {
	switch e {
	case Degraded:
		return true
	case Ok:
		return true
	default:
		return false
	}
}

// ClipAvailabilityResponse defines model for ClipAvailabilityResponse.
type ClipAvailabilityResponse struct {
	// Available True when the slug passes validation and has no live clip
	Available bool   `json:"available"`
	Slug      string `json:"slug"`
}

// CreateClipRequest defines model for CreateClipRequest.
type CreateClipRequest struct {
	// Body UTF-8 text body (256 KiB max)
	Body string `json:"body"`

	// Slug Case-sensitive clip name
	Slug string `json:"slug"`

	// TtlSeconds Chosen lifetime in seconds from the anonymous preset set (10 minutes,
	// 1 hour, or 2 hours). Omitted values use the default preset (2 hours).
	TtlSeconds *CreateClipRequestTtlSeconds `json:"ttlSeconds,omitempty"`
}

// CreateClipRequestTtlSeconds Chosen lifetime in seconds from the anonymous preset set (10 minutes,
// 1 hour, or 2 hours). Omitted values use the default preset (2 hours).
type CreateClipRequestTtlSeconds int64

// CreateClipResponse defines model for CreateClipResponse.
type CreateClipResponse struct {
	// ExpiresAt Expiry anchored to creation time as RFC3339Nano UTC
	ExpiresAt time.Time `json:"expiresAt"`
	Slug      string    `json:"slug"`
}

// DatabaseHealth defines model for DatabaseHealth.
type DatabaseHealth struct {
	// LatencyMs Round-trip latency in milliseconds when reachable
	LatencyMs *int64 `json:"latencyMs,omitempty"`
	Reachable bool   `json:"reachable"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Code Machine-readable error code
	Code string `json:"code"`

	// Message Human-readable message
	Message string `json:"message"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Database DatabaseHealth `json:"database"`

	// Migrations Present only when the pending-migration check completed
	Migrations *MigrationHealth `json:"migrations,omitempty"`

	// ServerTime Postgres `now()` as RFC3339Nano UTC, when reachable
	ServerTime *time.Time `json:"serverTime,omitempty"`

	// Status Overall status derived from the checks that completed
	Status HealthResponseStatus `json:"status"`
}

// HealthResponseStatus Overall status derived from the checks that completed
type HealthResponseStatus string

// MigrationHealth defines model for MigrationHealth.
type MigrationHealth struct {
	CurrentVersion int64 `json:"currentVersion"`
	Pending        bool  `json:"pending"`
}

// ReadClipResponse defines model for ReadClipResponse.
type ReadClipResponse struct {
	Body      string    `json:"body"`
	ExpiresAt time.Time `json:"expiresAt"`

	// ServerTime Server clock at response time as RFC3339Nano UTC
	ServerTime time.Time `json:"serverTime"`
	Slug       string    `json:"slug"`
}

// MethodNotAllowed defines model for MethodNotAllowed.
type MethodNotAllowed = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// RateLimited defines model for RateLimited.
type RateLimited = ErrorResponse

// ValidationError defines model for ValidationError.
type ValidationError = ErrorResponse

// CreateClipJSONRequestBody defines body for CreateClip for application/json ContentType.
type CreateClipJSONRequestBody = CreateClipRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// CreateClip Create a clip
	// (POST /api/clips)
	CreateClip(w http.ResponseWriter, r *http.Request)
	// GetClip Read a live clip
	// (GET /api/clips/{slug})
	GetClip(w http.ResponseWriter, r *http.Request, slug string)
	// GetClipAvailability Check whether a slug looks claimable
	// (GET /api/clips/{slug}/availability)
	GetClipAvailability(w http.ResponseWriter, r *http.Request, slug string)
	// GetHealth Stack health
	// (GET /api/health)
	GetHealth(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// CreateClip operation middleware
func (siw *ServerInterfaceWrapper) CreateClip(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateClip(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetClip operation middleware
func (siw *ServerInterfaceWrapper) GetClip(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "slug" -------------
	var slug string

	err = runtime.BindStyledParameterWithOptions("simple", "slug", r.PathValue("slug"), &slug, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "slug", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetClip(w, r, slug)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetClipAvailability operation middleware
func (siw *ServerInterfaceWrapper) GetClipAvailability(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "slug" -------------
	var slug string

	err = runtime.BindStyledParameterWithOptions("simple", "slug", r.PathValue("slug"), &slug, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "slug", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetClipAvailability(w, r, slug)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/health", wrapper.GetHealth)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/clips", wrapper.CreateClip)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/clips/{slug}", wrapper.GetClip)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/clips/{slug}/availability", wrapper.GetClipAvailability)

	return m
}